requirements.txt
click==7.1.2
configparser==5.0.0
Flask==1.1.2
Flask-FlatPages==0.7.2
Flask-Language==0.1.0
Flask-MySQL==1.5.1
Frozen-Flask==0.15
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
numpy==1.18.4
pandas==1.0.3
protobuf==3.12.2
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
Werkzeug==1.0.1
